==> tornado/notebooks/02. get_h3_hexagons.ipynb <==
import pandas as pd, geopandas as gpd, h3, shapely
import urllib, time, requests, psycopg2, urllib.parse, glob, os, us
from sqlalchemy import create_engine  
from h3 import h3
from shapely.ops import unary_union, cascaded_union
from shapely.geometry import mapping, Polygon

from concurrent.futures import ThreadPoolExecutor
from tqdm import tqdm

from pathlib import Path

import warnings
warnings.filterwarnings("ignore")

### 1. Generate US state mapping and geoDataframe

def generate_state_abbreviations():
    return {state.name: state.abbr for state in us.STATES_AND_TERRITORIES}

state_abbreviations = generate_state_abbreviations()
print(state_abbreviations)

# Load US states GeoDataFrame
us_states = gpd.read_file("https://eric.clst.org/assets/wiki/uploads/Stuff/gz_2010_us_040_00_5m.json")
us_states['area'] = us_states.geometry.area
us_states = us_states.sort_values(by='area', ascending=True)
us_states['st'] = us_states['NAME'].map(state_abbreviations)
us_states.head(5)

# Filter data to include only rows where the state is Arkansas
state = us_states[us_states['NAME'] == 'Arkansas']
state.head()

# Assuming 'arkansas_row' is the row for Arkansas from your original DataFrame
state_name = state['NAME'].values[0]

### 2. Generate H3 by State -- Set the Resolution: 1 - 15

def h3_id_to_geometry(h3_id):
    """
    Convert an H3 ID to a Shapely geometry.
    :param h3_id: The H3 ID
    :return: A Shapely Polygon representing the H3 hexagon
    """
    geo_boundary = h3.h3_to_geo_boundary(h3_id, geo_json=True)
    polygon = Polygon(geo_boundary)
    return polygon

def generate_h3_hexagons(state_geometry, resolution=resolution):
    """
    Generate H3 hexagons for a given state geometry.
    :param state_geometry: A Shapely geometry representing the state
    :param resolution: The resolution of the H3 hexagons
    :return: A DataFrame with H3 hexagons and their geometries
    """
    polygons = []

    if state_geometry.geom_type == 'Polygon':
        polygons.append(state_geometry)
    elif state_geometry.geom_type == 'MultiPolygon':
        for polygon in state_geometry:
            polygons.append(polygon)
    else:
        raise ValueError('Invalid geometry type. Only Polygon and MultiPolygon are supported.')

    hexagons = set()

    for polygon in polygons:
        # Get the bounding box of the polygon
        min_lng, min_lat, max_lng, max_lat = polygon.bounds

        # Convert the polygon to GeoJSON format
        polygon_geojson = gpd.GeoSeries([polygon]).__geo_interface__["features"][0]["geometry"]

        # Generate a list of H3 hexagons within the bounding box
        hexagons_in_bbox = h3.polyfill(
            polygon_geojson,
            res=resolution,
            geo_json_conformant=True
        )

        # Filter out hexagons that are not within the polygon
        for hexagon in hexagons_in_bbox:
            hexagon_boundary = Polygon(h3.h3_to_geo_boundary(hexagon, geo_json=True))
            if polygon.contains(hexagon_boundary):
                hexagons.add(hexagon)

    # Convert the list of hexagons into a DataFrame
    hexagon_df = pd.DataFrame(list(hexagons), columns=['h3_hexagon'])

    # Add the geometry column to the DataFrame
    hexagon_df['geometry'] = hexagon_df['h3_hexagon'].apply(h3_id_to_geometry)

    return hexagon_df


resolution = 4

# Generate the H3 hexagons
gdf = generate_h3_hexagons(state['geometry'].values[0])
gdf = gpd.GeoDataFrame(gdf)

# Call the state name
state_name = state['NAME'].values[0]

# Add the state name to the output DataFrame
gdf['state'] = state_name


gdf.head(10)

gdf.to_file(f"data/h3_{state_name}_{resolution}.gpkg", driver="GPKG")
