==> tornado/notebooks/03. dissolve_hexagons.ipynb <==
import pandas as pd
import geopandas as gpd
import os, us, warnings, datetime
from sqlalchemy import create_engine  
from datetime import datetime
from pathlib import Path
warnings.filterwarnings("ignore")

### Prepare US state geo data

def generate_state_abbreviations():
    return {state.name: state.abbr for state in us.STATES_AND_TERRITORIES}

state_abbreviations = generate_state_abbreviations()
print(state_abbreviations)

# Load US states GeoDataFrame
us_states = gpd.read_file("https://eric.clst.org/assets/wiki/uploads/Stuff/gz_2010_us_040_00_5m.json")
us_states['area'] = us_states.geometry.area
us_states = us_states.sort_values(by='area', ascending=True)
us_states['st'] = us_states['NAME'].map(state_abbreviations)
us_states.head(5)

### For each US state, spatial join and aggregate Tornado data into H3 hexagons

def calculate_years_diff(dates_list):
    if len(dates_list) > 1:
        years_diff = [(dates_list[i] - dates_list[i - 1]).days / 365.25 for i in range(1, len(dates_list))]
        return [round(year_diff, 3) for year_diff in years_diff]
    else:
        return []

def calculate_days_diff(dates):
    if len(dates) > 1:
        sorted_dates = sorted(dates)
        return [(sorted_dates[i + 1] - sorted_dates[i]).days for i in range(len(sorted_dates) - 1)]
    else:
        return []

from datetime import datetime

def convert_to_date(dates_list):
    return [datetime.strptime(date_str, '%Y-%m-%d').date() for date_str in dates_list]

# Define a function that calculates the years between the event date and the present date
def years_since_event(dates):
    if len(dates) == 1:
        # if there is only one event, calculate the years since that event
        return (datetime.now().date() - dates[0]).days / 365.25
    else:
        # if there are multiple events, return None (or whatever you want to return in this case)
        return None

def load_geodata(state_name, directory):
    print(f"Read in geoDataframe for : {state_name}")
    st_boundary = us_states[us_states['st'] == state_name]
    st_boundary = st_boundary.to_crs(4326)
    print(f"Loading H3 id's and polygons for: {state_name}")
    try:
        full_name = st_boundary['NAME'].iloc[0]
    except IndexError:
        raise Exception("IndexError: The DataFrame st_boundary is likely empty or the index is out of bounds.")

    print("We've created the two state objects: full_name and abbreviated name")
    hexagons = gpd.read_file(f"{directory}/h3_by_state/h3_{full_name}.gpkg")
    hexagons['state'] = state_name
    lookup = hexagons['state'][0]
    st_boundary = st_boundary.dropna()

    return hexagons, lookup, st_boundary

def load_event_data(directory, lookup, st_boundary):
    event = gpd.read_file(f"{directory}/input/tornado_tracks.shp")
    event = event.to_crs(4326)
    event_df = event[event['st'] == lookup]
    print("Created event geoDataframe")
    event_df = gpd.clip(event_df, st_boundary)
    return event_df

def process_event_history(s):
    # Regular expression pattern to match the dates
    pattern = r'datetime\.date\((\d{4}), (\d{1,2}), (\d{1,2})\)'

    # Extract all matching date components
    matches = re.findall(pattern, s)

    # Convert to formatted date strings
    formatted_dates = [datetime(int(year), int(month), int(day)).strftime('%m/%d/%Y') for year, month, day in matches]

    return formatted_dates

print(us_states['st'].unique())


# Set directory path
directory = Path("/tornado/data")
output_directory = "h3_hexagons_geopackage"

# Loop through each state name
for state_name in us_states['st'].dropna().unique():
    try:
        print("________________________________________________________________")
        hexagons, lookup, st_boundary = load_geodata(state_name, directory)
        event_df = load_event_data(directory, lookup, st_boundary)

        # Rest of the processing steps would go here...
        # Clip the data using GeoPandas clip
        event_df = gpd.clip(event_df, st_boundary)
        event_df['buffer_geom'] = event_df.buffer(0.008)
        event_df = event_df.drop('geometry', axis=1)
        event_df = event_df.rename(columns={'buffer_geom': 'geometry'})
        print(f"Loading Tornado event layer by the state of: {lookup}")

        # Convert column A from float to string and remove ".0"
        event_df['yr'] = event_df['yr'].astype(int).astype(str).str.replace('\.0', '')
        event_df['mo'] = event_df['mo'].astype(int).astype(str).str.replace('\.0', '')
        event_df['dy'] = event_df['dy'].astype(int).astype(str).str.replace('\.0', '')

                # select only float columns
        float_cols = event_df.select_dtypes(include=['float'])

        # round float columns to 2 decimal places
        event_df[float_cols.columns] = float_cols.round(2)

        #event_df['date'] = pd.to_datetime(event_df['date'], errors='coerce').dt.date
        event_df = event_df[['geometry', 'yr', 'mo', 'dy', 'date', 'time', 'tz', 'st', 'mag', 'inj',
        'fat', 'loss', 'closs', 'len', 'wid']]
        print("Start Join of geoDataframe")

        # Use geopandas to spatial join *intersect* the two tables
        join_df = gpd.sjoin(hexagons, event_df, how='inner', predicate='intersects')
        print(f"Resulted join has a record count of: {len(join_df.index)}")

        # Sort and group the data to produce aggregate layer
        join_df = join_df.sort_values(by=['h3_hexagon', 'date'], ascending=True)
        group_df = join_df.groupby('h3_hexagon').agg({
            'yr': list, 
            'mo': list, 
            'date': list, 
            'mag': list, 
            'loss': list, 
            'closs': list, 
            'inj': list, 
            'fat': list, 
            'len': list, 
            'wid': list
            })
        print("Aggregating event geoDataframe")

        # Convert the strings in the 'date' column to datetime.date objects
        group_df['date'] = group_df['date'].apply(convert_to_date)

        # Calculate days difference between dates in the 'date' column
        group_df['days_diff'] = group_df['date'].apply(calculate_days_diff)
        group_df['years_diff'] = group_df['date'].apply(calculate_years_diff)

        # Calculate total number of tornado events and sum of years difference
        group_df['num_events'] = group_df['date'].apply(len)
        group_df['sum_years_diff'] = group_df['years_diff'].apply(sum)
        group_df['sum_property_loss'] = group_df['loss'].apply(sum)

        # Calculate the average loss per year
        group_df['total_loss'] = group_df['loss'].apply(sum)
        group_df['total_years'] = group_df['date'].apply(len)
        group_df['avg_loss_per_year'] = group_df['total_loss'] / group_df['total_years']

        ######## Addition
        # Apply the function to the 'date' column
        group_df['years_since_event'] = group_df['date'].apply(years_since_event)

        # replace zero or null values with np.nan
        group_df['sum_years_diff'] = group_df['sum_years_diff'].replace({0: np.nan, None: np.nan})

        # Now use 'years_since_event' where 'sum_years_diff' is null
        group_df['sum_years_diff'] = group_df['sum_years_diff'].where(group_df['sum_years_diff'].notnull(), group_df['years_since_event'])

        # Now calculate average frequency as before
        group_df['avg_frequency'] = group_df['num_events'] / group_df['sum_years_diff']
        ######## End Addition

        # Calculate average frequency
        group_df['avg_property_loss'] = group_df['sum_property_loss'] / group_df['num_events']
        # Calculate average frequency of mag_stats#
        group_df['avg_mag_freq'] = group_df['num_events'] / group_df['sum_years_diff']

        print("Dissolving event geoDataframe")
        # Run a Group By and Sum operation to produce aggregate layer
        dissolve_df = join_df.dissolve(
                by="h3_hexagon",
                aggfunc={'state': 'count',
                        'mag':'mean', 'inj': 'sum', 
                        'fat': 'sum', 'loss': 'sum', 
                        'closs': 'sum', 'len': 'mean', 'wid': 'mean'})

        dissolve_df = dissolve_df.rename(columns={'h3_hexagon':'h3_id', 'name':'tornado_count'})

        # Merge the two dataframes to produce final aggregate layer
        final = dissolve_df.join(group_df, lsuffix='_h3_history', rsuffix='_stats')
        final = final.rename(columns={'state': 'tornado_count'})

        df_melt = final.assign(names=final.date.str.split(","))
                
        final = df_melt.date.apply(pd.Series) \
            .merge(df_melt, right_index=True, left_index=True)

        final['date'] = final['date'].apply(convert_to_string)


        ##############################

        print("Successfully dissolved...")
        final.to_csv(f"{directory}/h3_hexagons_tornado/{lookup}_final_geom.csv")
        print("Wrote to csv...")
    except FileNotFoundError as e:
        print(f"Error processing {state_name}: {e}")
    except ValueError as e:
        print(f"Error processing {state_name}: {e}")

## Step-by-Step Model: DO NOT DELETE

from pathlib import Path
import glob

# Set directory path
directory = Path("/tornado/data")
output_directory = "h3_hexagons_geopackage"


# Specify the state you want to process
state_name = 'AR'

hexagons, lookup, st_boundary = load_geodata(state_name, directory)
event_df = load_event_data(directory, lookup, st_boundary)

# Rest of the processing steps would go here...
# Clip the data using GeoPandas clip
event_df = gpd.clip(event_df, st_boundary)
event_df['buffer_geom'] = event_df.buffer(0.008)
event_df = event_df.drop('geometry', axis=1)
event_df = event_df.rename(columns={'buffer_geom': 'geometry'})
print(f"Loading Tornado event layer by the state of: {lookup}")

event_df.head(3)

# Convert column A from float to string and remove ".0"
event_df['yr'] = event_df['yr'].astype(int).astype(str).str.replace('\.0', '')
event_df['mo'] = event_df['mo'].astype(int).astype(str).str.replace('\.0', '')
event_df['dy'] = event_df['dy'].astype(int).astype(str).str.replace('\.0', '')

event_df.head(3)

# select only float columns
float_cols = event_df.select_dtypes(include=['float'])

# round float columns to 2 decimal places
event_df[float_cols.columns] = float_cols.round(2)

#event_df['date'] = pd.to_datetime(event_df['date'], errors='coerce').dt.date
event_df = event_df[['geometry', 'yr', 'mo', 'dy', 'date', 'time', 'tz', 'st', 'mag', 'inj',
'fat', 'loss', 'closs', 'len', 'wid']]
print("Start Join of geoDataframe")

# Use geopandas to spatial join *intersect* the two tables
join_df = gpd.sjoin(hexagons, event_df, how='inner', predicate='intersects')
print(f"Resulted join has a record count of: {len(join_df.index)}")

# Sort and group the data to produce aggregate layer
join_df = join_df.sort_values(by=['h3_hexagon', 'date'], ascending=True)

join_df.head(5)

group_df = join_df.groupby('h3_hexagon').agg({
    'yr': list, 
    'mo': list, 
    'date': list, 
    'mag': list, 
    'loss': list, 
    'closs': list, 
    'inj': list, 
    'fat': list, 
    'len': list, 
    'wid': list
    })
print("Aggregating event geoDataframe")

group_df.head(5)

import datetime

# Convert the strings in the 'date' column to datetime.date objects
group_df['date'] = group_df['date'].apply(convert_to_date)

# Calculate days difference between dates in the 'date' column
group_df['days_diff']_df['date'].apply(calculate_days_diff)

group_df['years_diff'] = group_df['date'].apply(calculate_years_diff)

group_df.head(5)

# Calculate total number of tornado events and sum of years difference
group_df['num_events'] = group_df['date'].apply(len)
group_df['sum_years_diff'] = group_df['years_diff'].apply(sum)
group_df['sum_property_loss'] = group_df['loss'].apply(sum)

group_df.head(5)

# Calculate the average loss per year
group_df['total_loss'] = group_df['loss'].apply(sum)
group_df['total_years'] = group_df['date'].apply(len)
group_df['avg_loss_per_year'] = group_df['total_loss'] / group_df['total_years']

group_df.head(5)

from datetime import datetime


# replace zero or null values with np.nan
group_df['sum_years_diff'] = group_df['sum_years_diff'].replace({0: np.nan, None: np.nan})

# Apply the function to the 'date' column
group_df['years_since_event'] = group_df['date'].apply(years_since_event)

# Now use 'years_since_event' where 'sum_years_diff' is null
group_df['sum_years_diff'] = group_df['sum_years_diff'].where(group_df['sum_years_diff'].notnull(), group_df['years_since_event'])

# Now calculate average frequency as before
group_df['avg_frequency'] = group_df['num_events'] / group_df['sum_years_diff']

group_df.head(5)

group_df['avg_property_loss'] = group_df['sum_property_loss'] / group_df['num_events']
# Calculate average frequency of mag_stats
group_df['avg_mag_freq'] = group_df['num_events'] / group_df['sum_years_diff']

group_df.head(5)

print("Dissolving event geoDataframe")
# Run a Group By and Sum operation to produce aggregate layer
dissolve_df = join_df.dissolve(
        by="h3_hexagon",
        aggfunc={'state': 'count',
                'mag':'mean', 'inj': 'sum', 
                'fat': 'sum', 'loss': 'sum', 
                'closs': 'sum', 'len': 'mean', 'wid': 'mean'})


dissolve_df = dissolve_df.rename(columns={'h3_hexagon':'h3_id', 'name':'tornado_count'})

dissolve_df.head(3)

# Merge the two dataframes to produce final aggregate layer
final = dissolve_df.join(group_df, lsuffix='_h3_history', rsuffix='_stats')
final = final.rename(columns={'state': 'tornado_count'})

final.head(5)

df_melt = final.assign(names=final.date.str.split(","))

df_melt.head(5)

final = df_melt.date.apply(pd.Series) \
    .merge(df_melt, right_index=True, left_index=True)

final.head(5)

def convert_to_string(dates_list):
    return [date.strftime('%Y-%m-%d') for date in dates_list]

final['date'] = final['date'].apply(convert_to_string)


final.columns

#final_stats = final.drop([0, 1, 'geometry'], axis=1)
#final_geom = final.drop([0], axis=1)

##############################

print("Successfully dissolved...")
final.to_csv(directory / f"h3_hexagons_tornado/{lookup}_final_geom.csv")
print("Wrote to csv...")
