==> tornado/notebooks/05. tornado_stats_h3_merge.ipynb <==
from pathlib import Path   
import geopandas as gpd
import pandas as pd
from pandas import json_normalize
from shapely import wkt
from datetime import datetime, date, timedelta
import numpy as np
import re
from pathlib import Path
from scipy import stats

import warnings
warnings.filterwarnings("ignore")

from io import StringIO
import matplotlib.pyplot as plt
from urllib.parse import urlparse
from sqlalchemy import create_engine

## Load Data

folder = Path('/tornado/data/h3_hexagons_tornado/')

gdf = pd.concat([
    pd.read_csv(shp).assign(state=shp.stem.split("_")[0])
    for shp in folder.glob("*_final_geom.csv")
]).pipe(gpd.GeoDataFrame)

gdf = gdf.drop(columns=['0','1','2','3','4','5','6','names', '7', '8', '9', '10',
       '11', '12', '13', '14', '15', '16', '17', '18', '19', '20','yr','total_years'])

gdf = gdf.rename(columns={'h3_hexagon':'h3_id',
 'mag_h3_history':'magnitude_mean',
 'inj_h3_history':'injuries_sum',
 'fat_h3_history':'fatality_sum',
 'loss_h3_history':'property_loss_sum',
 'closs_h3_history':'crop_loss_sum',
 'len_h3_history':'t_length_history',
 'wid_h3_history':'t_width_history',
 'mo':'month',
 'date':'event_history',
 'mag_stats':'magnitude_stats',
 'loss_stats':'property_loss_stats',
 'closs_stats':'crop_loss_stats',
 'inj_stats':'injuries_stats',
 'fat_stats':'fatalities_stats',
 'len_stats':'length_stats',
 'wid_stats':'width_stats'})

gdf.head(2)

import geopandas as gpd
from shapely import wkt

# Convert the geometry column to WKT
gdf['geometry'] = gdf['geometry'].apply(lambda wkt_string: wkt.loads(wkt_string))

gdf = gpd.GeoDataFrame(gdf, geometry='geometry', crs=4326)

gdf = gdf[['h3_id', 'geometry', 'tornado_count', 'magnitude_mean', 'injuries_sum',
       'fatality_sum', 'property_loss_sum', 'crop_loss_sum','month', 'event_history',
       'magnitude_stats', 'property_loss_stats', 'crop_loss_stats',
       'injuries_stats', 'fatalities_stats', 'length_stats', 'width_stats',
       'days_diff', 'years_diff', 'sum_years_diff',
       'sum_property_loss', 'total_loss', 'avg_loss_per_year', 'avg_frequency',
       'avg_property_loss', 'avg_mag_freq', 'state']]

gdf1 = gdf

gdf = gdf1

## Start Data Inspection Stage

import numpy as np
from scipy import stats

def inspect_data(gdf):
    inspection_results = {}
    for col in gdf.select_dtypes(include=[np.number]).columns:
        # Calculate IQR
        Q1 = gdf[col].quantile(0.25)
        Q3 = gdf[col].quantile(0.75)
        IQR = Q3 - Q1

        # Identify outliers
        is_outlier = ~((gdf[col] >= (Q1 - 1.5 * IQR)) & (gdf[col] <= (Q3 + 1.5 * IQR)))
        num_outliers = is_outlier.sum()

        # Calculate stats
        stats_dict = {
            "data_type": gdf[col].dtype,
            "min": gdf[col].min(),
            "median": gdf[col].median(),
            "max": gdf[col].max(),
            "mode": gdf[col].mode(),
            "num_outliers": num_outliers,
            "outlier_percentage": (num_outliers / len(gdf) * 100)
        }

        # Print results
        print(f'Column {col} has {num_outliers} extreme outliers out of {len(gdf)} rows ({stats_dict["outlier_percentage"]:.2f}%).')

        # Add to results
        inspection_results[col] = stats_dict
    return inspection_results



def validate_data(gdf, inspection_results):
    print("Validation Results:")
    for col, stats_dict in inspection_results.items():
        # Handle negative values
        if col in ["tornado_count", "injuries_sum", "fatality_sum", "property_loss_sum", "crop_loss_sum", "num_events", "sum_years_diff", "sum_property_loss", "total_loss", "avg_loss_per_year", "avg_frequency", "avg_property_loss"]:
            num_negative = (gdf[col] < 0).sum()
            negative_percentage = num_negative / len(gdf) * 100
            print(f"Column {col}: {num_negative} negative values ({negative_percentage:.2f}%)")

        # Add more validation checks as needed

    # Perform other validation checks and print the results

    return gdf


def cleanse_data(gdf, inspection_results):
    print("Data Cleansing:")
    for col, stats_dict in inspection_results.items():
        # Handle outliers
        if stats_dict["num_outliers"] > 0:
            print(f"Column {col}: {stats_dict['num_outliers']} outliers corrected")

        # Handle negative values
        if col in ["tornado_count", "injuries_sum", "fatality_sum", "property_loss_sum", "crop_loss_sum", "num_events", "sum_years_diff", "sum_property_loss", "total_loss", "avg_loss_per_year", "avg_frequency", "avg_property_loss"]:
            num_negative = (gdf[col] < 0).sum()
            print(f"Column {col}: {num_negative} negative values corrected")

        # Add more data cleansing steps as needed

    # Perform other data cleansing steps

    return gdf  # Return the updated geodataframe


from sklearn.preprocessing import MinMaxScaler

def normalize_data(gdf):
    # Create a scaler object
    scaler = MinMaxScaler()

    # Replace infinity values with NaN
    gdf = gdf.replace([np.inf, -np.inf], np.nan)

    # If there are any NaNs in the column, fill them with the column's mean
    for col in gdf.columns:
        if gdf[col].isna().sum() > 0: 
            gdf[col].fillna(gdf[col].mean(), inplace=True)

    # Now apply the scaler to each column
    for col in gdf.select_dtypes(include=[np.number]).columns:
        gdf[col] = scaler.fit_transform(gdf[[col]])

    return gdf

    # Perform data normalization and print the results


def process_data(gdf):
    # Call inspect_data and store the results
    inspection_results = inspect_data(gdf)

    # Validate the data using the inspection results
    validate_data(gdf, inspection_results)

    # Cleanse the data using the inspection results
    gdf = cleanse_data(gdf, inspection_results)

    # Normalize the data
    gdf = normalize_data(gdf)

    return gdf

# Call process_data to get the processed geodataframe
processed_gdf = process_data(gdf)


processed_gdf.head(5)

gdf.head(10)

def replace_inf(df, col):
    df[col] = np.where(df[col] == np.inf, 0, df[col])
    return df

gdf = replace_inf(gdf, 'avg_frequency')

# select only float columns
float_cols = gdf.select_dtypes(include=['float'])

# round float columns to 2 decimal places
gdf[float_cols.columns] = float_cols.round(2)

def process_event_history(s):
    # Regular expression pattern to match the dates
    pattern = r'datetime\.date\((\d{4}), (\d{1,2}), (\d{1,2})\)'

    # Extract all matching date components
    matches = re.findall(pattern, s)

    # Convert to formatted date strings
    formatted_dates = [datetime(int(year), int(month), int(day)).strftime('%m/%d/%Y') for year, month, day in matches]

    return formatted_dates

# Apply the function to the event_history column
gdf['event_history'] = gdf['event_history'].apply(process_event_history)

gdf['count_rank'] = (gdf['tornado_count'] - gdf['tornado_count'].min()) / (gdf['tornado_count'].max() - gdf['tornado_count'].min()) * 100

gdf['magnitude_rank'] = (gdf['magnitude_mean'] - gdf['magnitude_mean'].min()) / (gdf['magnitude_mean'].max() - gdf['magnitude_mean'].min()) * 100

gdf['injuries_rank'] = (gdf['injuries_sum'] - gdf['injuries_sum'].min()) / (gdf['injuries_sum'].max() - gdf['injuries_sum'].min()) * 100

gdf['fatalities_rank'] = (gdf['fatality_sum'] - gdf['fatality_sum'].min()) / (gdf['fatality_sum'].max() - gdf['fatality_sum'].min()) * 100

gdf['ploss_rank'] = (gdf['property_loss_sum'] - gdf['property_loss_sum'].min()) / (gdf['property_loss_sum'].max() - gdf['property_loss_sum'].min()) * 100

gdf['freq_rank'] = (gdf['avg_frequency'] - gdf['avg_frequency'].min()) / (gdf['avg_frequency'].max() - gdf['avg_frequency'].min()) * 100

gdf.head(5)

## Write to Postgres

connection_string = f'postgresql://:NJ2tFO6DndyH@ep-misty-pond-968466.us-east-2.aws.neon.tech/neondb'

url = urlparse(connection_string)

db_host = url.hostname
db_port = url.port
db_name = url.path[1:]
db_user = url.username
db_password = url.password

engine = create_engine(connection_string)

gdf.columns

gdf = gdf[['h3_id', 'geometry', 'tornado_count', 'magnitude_mean', 'injuries_sum',
       'fatality_sum', 'property_loss_sum', 'crop_loss_sum', 'month',
       'event_history', 'magnitude_stats', 'property_loss_stats',
       'crop_loss_stats', 'injuries_stats', 'fatalities_stats', 'days_diff', 'years_diff', 'sum_years_diff',
       'sum_property_loss', 'total_loss', 'avg_loss_per_year', 'avg_frequency', 'count_rank',
       'magnitude_rank', 'injuries_rank', 'fatalities_rank', 'ploss_rank',
       'freq_rank','state']]

table_name = f"tornado_final_stats"

# Write the GeoDataFrame to the database using to_postgis function
gdf.to_postgis(table_name, engine, if_exists='replace')
