==> tornado/notebooks/07. write_to_postgresql.ipynb <==
import os, requests, time, json, glob
from io import StringIO
from datetime import date, timedelta 
import matplotlib.pyplot as plt

import urllib.parse
from urllib.parse import urlparse

import pandas as pd
from pandas import json_normalize
import geopandas as gpd

import psycopg2
from sqlalchemy import create_engine, Column, Integer, Float, String, DateTime

from geoalchemy2 import Geometry, WKTElement

import warnings
warnings.filterwarnings("ignore")

### Set the instance PostgreSQL to Neon.Tech

connection_string = f''

url = urlparse(connection_string)

db_host = url.hostname
db_port = url.port
db_name = url.path[1:]
db_user = url.username
db_password = url.password

def connect_to_db():
    connection = None
    try:
        connection = psycopg2.connect(
            host=db_host,
            port=db_port,
            dbname=db_name,
            user=db_user,
            password=db_password
        )
        print("Connected to the PostgreSQL instance!")
    except (Exception, psycopg2.DatabaseError) as error:
        print("Error while connecting to PostgreSQL", error)

    return connection

def execute_query(query):
    connection = connect_to_db()

    if connection is not None:
        try:
            cursor = connection.cursor()
            cursor.execute(query)
            result = cursor.fetchall()
            print("Query result:", result)
        except (Exception, psycopg2.DatabaseError) as error:
            print("Error while executing query", error)
        finally:
            if connection is not None:
                connection.close()
                print("Connection closed.")

connection = connect_to_db()

### Our connection to the database has been made

# Example query to get all rows from a table named "your_table_name"
example_query = "SELECT * FROM geodata;"
execute_query(example_query)

### Configure the PostgreSQL instance to handle PostGIS + other extensions

def enable_postgis_extensions(connection):
    try:
        cursor = connection.cursor()
        cursor.execute("CREATE EXTENSION IF NOT EXISTS postgis;")
        cursor.execute("CREATE EXTENSION IF NOT EXISTS postgis_topology;")
        cursor.execute("CREATE EXTENSION IF NOT EXISTS fuzzystrmatch;")
        cursor.execute("CREATE EXTENSION IF NOT EXISTS postgis_tiger_geocoder;")
        connection.commit()
        print("PostGIS-related extensions enabled.")
    except (Exception, psycopg2.DatabaseError) as error:
        print("Error while enabling PostGIS-related extensions", error)

connection = connect_to_db()
if connection is not None:
    enable_postgis_extensions(connection)
    connection.close()


### Write geoDataframe to PostgreSQL

entity_name = "tornado_tracks"

# Read your GeoDataFrame from a file or create it programmatically
gdf = gpd.read_file(f"data/{entity_name}.shp")

gdf = gpd.read_file(f"data/tornado_stats_final.csv")

gdf.info()

cursor = connection.cursor()

query = '''
SELECT *
FROM tornado_tracks
'''

cursor.execute(query)
data = cursor.fetchall()

# Get the column names from the cursor description
columns = [desc[0] for desc in cursor.description]

# Create a DataFrame from the fetched data and column names
df = pd.DataFrame(data, columns=columns)

# Get the number of rows and columns in the DataFrame
print(f"The DataFrame has {df.shape[0]} rows and {df.shape[1]} columns")

# Calculate the mean value of the 'length' column
print(f"The mean length of tornado tracks is {df['len'].mean()} miles")

# Show the first 10 rows of the DataFrame
df.head(10)


# define SQL query
query = "SELECT *,  ST_AsText(geometry) AS wkt_geom FROM tornado_tracks;"

# read data into GeoDataFrame
gdf = gpd.read_postgis(query, connection, geom_col="wkt_geom")

# convert wkt_geom column from string to geometry type
gdf["wkt_geom"] = gdf["wkt_geom"].apply(lambda x: shapely.wkt.loads(x))


connection = connect_to_db()
cursor = connection.cursor()

query = '''
SELECT table_name
FROM information_schema.tables
WHERE table_schema = 'public'
'''

cursor.execute(query)
tables = cursor.fetchall()

for table in tables:
    print(table[0])


### Function: Fix Invalid Geometries (pre-processing)

import geopandas as gpd

# Test for invalid geometries in a GeoDataFrame
invalid_geom = gdf[~gdf.is_valid]
if not invalid_geom.empty:
    print('Invalid geometries found')

def fix_invalid_geometries(gdf, buffer_distance=0.0001):
    # Find invalid geometries
    invalid_geom = gdf[~gdf.is_valid]
    if invalid_geom.empty:
        return gdf

    # Fix invalid geometries by buffering with a small distance
    gdf.loc[~gdf.is_valid, 'geometry'] = invalid_geom.buffer(buffer_distance)

    # Re-test for invalid geometries and recursively fix them if necessary
    return fix_invalid_geometries(gdf)

fixed_gdf = fix_invalid_geometries(gdf)


import geopandas as gpd
import dask.dataframe as dd
import dask_geopandas as dask_gpd
from shapely.geometry import shape, mapping
#from shapely.ops import make_valid


def fix_invalid_geometry(geometry):
    if geometry.is_valid:
        return geometry
    else:
        return geometry.buffer(0.0001)

# Read the data with GeoPandas
# Convert GeoPandas GeoDataFrame to Dask-GeoPandas GeoDataFrame
dgdf = dask_gpd.from_geopandas(gdf, npartitions=4)

# Apply the fix_invalid_geometry function in a distributed way using Dask
dgdf['geometry'] = dgdf['geometry'].map_partitions(lambda part: part.apply(fix_invalid_geometry), meta=('geometry', 'object'))

dgdf['geometry']

# Compute the results
fixed_geometries_gdf = dgdf.compute()

# Save the fixed geometries to a new file
#fixed_geometries_gdf.to_file("fixed_geometries.geojson", driver="GeoJSON")
