==> tornado/notebooks/08. real_estate_impact.ipynb <==
import os, requests, time, json, glob
from io import StringIO
from datetime import date
from datetime import timedelta 
import matplotlib.pyplot as plt

import pandas as pd
import geopandas as gpd
#import h3pandas

from geoalchemy2 import Geometry, WKTElement
import psycopg2
from sqlalchemy import create_engine

from pandas import json_normalize
import urllib, urllib3, urllib.parse

df = pd.read_csv('datasets/fake_real_estate.csv')

df.head(5)

gdf = gpd.GeoDataFrame(
    df, geometry=gpd.points_from_xy(df.longitude, df.latitude))

gdf.head(5)

gdf.to_file('datasets/fake_real_estate.gpkg', driver='GPKG', layer='name')
